requirements.txt
pillow>=10.0.0
streamlit>=1.28.0
